Alterações dos metodos
Alteração no código inteiro.

+++ src/main/java/com/tecdes/pedido/controller/ProdutoController.java
@@ -1,15 +1,12 @@
 package com.tecdes.pedido.controller;
 
-
 import java.util.List;
-
 
 import com.tecdes.pedido.model.entity.Produto;
 import com.tecdes.pedido.service.ProdutoService;
 
 
 public class ProdutoController {
-
 
     private final ProdutoService service;
 
@@ -19,34 +16,27 @@ public class ProdutoController {
     }
 
 
-    // Salva
     public void save(String nome ,  double preco, String categoria, String descricao) {
         service.salvarProduto(nome, preco, categoria, descricao);
     }
 
 
-    // Busca todos
     public List<Produto> buscarTodos() {
         return service.buscarTodos();
     }
 
 
-     // Busca por Id
-     public Produto findById(int idProduto){
+    public Produto findById(Long idProduto){
         return service.buscarPorId(idProduto);
     }
 
 
-    // Atualiza
-    public void update(int idProduto, String nome ,  double preco, String categoria, String descricao){
+    public void update(Long idProduto, String nome ,  double preco, String categoria, String descricao){
         service.atualizarProduto(idProduto, nome, preco, categoria, descricao);
     }
 
 
-    // Deleta
-    public void delete(int idProduto){
+    public void delete(Long idProduto){
         service.deletarProduto(idProduto);
     }
 }
-
-
